Create Working full deleting AND archive replacement.txt
This Office scripts Removes any collumn which is not vital to a centra upload. So make sure to erase the text string you want to update in lines 16-102.

There is a secondary function in this code where it also checks if there is a "category" column. If there is, the code replaceses any "archive" collections with an empty string.